+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.5.1
 - Improve time performance by internally merging TypeScript into JavaScript.
 - Improve import generation recognition.
+- Add reading type definitions from index.d.ts and `types` in package.json.
 
 ## 1.5.0
 - Add ability to sort recently used items at the top of the list.